mv DOM-related Lib function to src/lib/dom.js
